refactor(scanner): optimize deduplicator fingerprint, enforce strict cycle assertion, and encapsulate scanner modules
