+++ src/scanner/mod.rs
@@ -1,7 +1,7 @@
 pub mod conditions;
 pub mod core;
-pub mod dataflow;
-pub mod flow_tracker;
+mod dataflow;
+mod flow_tracker;
 pub mod modes;
 mod multifile_taint;
 mod output;
@@ -9,7 +9,7 @@ mod parser_helper;
 pub mod prefilter;
 mod scan_context;
 mod scanning_logic;
-pub mod taint_utils;
+mod taint_utils;
 pub mod utils;
 mod vulnerability_scanner;
