+++ src/scanner/taint_utils.rs
@@ -1,28 +1,43 @@
-#![allow(clippy::too_many_arguments, clippy::large_enum_variant, clippy::needless_range_loop)]
+use std::collections::hash_map::DefaultHasher;
+use std::hash::{Hash, Hasher};
 
 use crate::common::CommonUtils;
 
 #[derive(Debug, Clone)]
-pub struct TaintRuleDeduplicator {
+pub(crate) struct TaintRuleDeduplicator {
     /// Mapping from (source_pattern, sink_pattern) to the rule that should handle it
     rule_mapping: std::collections::BTreeMap<(String, String), crate::rules::UnifiedRule>,
     /// Consolidated source patterns across all rules
     source_patterns: std::collections::BTreeSet<String>,
     /// Consolidated sink patterns across all rules
     pub(crate) sink_patterns: std::collections::BTreeSet<String>,
+    /// Precalculated u64 hash fingerprint identifying this deduplicator rule set
+    fingerprint: u64,
 }
 
 impl TaintRuleDeduplicator {
     /// Create a new deduplicator from a list of taint rules
-    pub fn new(taint_rules: &[&crate::rules::UnifiedRule]) -> Self {
+    pub(crate) fn new(taint_rules: &[&crate::rules::UnifiedRule]) -> Self {
+        let mut hasher = DefaultHasher::new();
+
         let mut deduplicator = Self {
             rule_mapping: std::collections::BTreeMap::new(),
             source_patterns: std::collections::BTreeSet::new(),
             sink_patterns: std::collections::BTreeSet::new(),
+            fingerprint: 0,
         };
 
         // Process each rule and create specific source-sink mappings
         for rule in taint_rules {
+            rule.id.hash(&mut hasher);
+            rule.mode.hash(&mut hasher);
+            if let Some(sources) = &rule.sources {
+                sources.hash(&mut hasher);
+            }
+            if let Some(sinks) = &rule.sinks {
+                sinks.hash(&mut hasher);
+            }
+
             if let (Some(sources), Some(sinks)) = (&rule.sources, &rule.sinks) {
                 // Add all patterns to consolidated sets
                 for source in sources {
@@ -42,7 +57,13 @@ impl TaintRuleDeduplicator {
             }
         }
 
+        deduplicator.fingerprint = hasher.finish();
         deduplicator
+    }
+
+    /// Return the precomputed rule set fingerprint
+    pub(crate) fn fingerprint(&self) -> u64 {
+        self.fingerprint
     }
 
     /// Get the specific rule for a source-sink combination
